+++ packages/router/test/adapters/contract/contract.spec.ts
@@ -528,7 +528,7 @@ describe("Contract Adapter", () => {
     });
   });
 
-  describe("fulfillRouterContract", async () => {
+  describe("#fulfillRouterContract", async () => {
     beforeEach(() => {
       sanitationStub = stub(SharedFns, "sanitationCheck").resolves();
     });